function_app/requirements.txt
azure-functions>=1.18.0
openai>=1.40.0
httpx>=0.27.0
pydantic>=2.6.0
